refactor(story5.2): delete lotto constant values and move to lotto class
