--- src/main/java/lotto/common/LottoConstants.java
@@ -1,7 +1,0 @@
-package lotto.common;
-
-public class LottoConstants {
-    public static final int LOTTO_NUMBER_MAX_COUNT = 6;
-    public static final int LOTTO_NUMBER_MIN = 1;
-    public static final int LOTTO_NUMBER_MAX = 45;
-}
